+++ backend/canisters/community/api/src/updates/update_channel.rs
@@ -16,6 +16,7 @@ pub struct Args {
     pub events_ttl: OptionUpdate<Milliseconds>,
     pub gate: OptionUpdate<AccessGate>,
     pub public: Option<bool>,
+    pub messages_visible_to_non_members: Option<bool>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]