+++ backend/canisters/group_index/impl/src/model/public_groups.rs
@@ -97,6 +97,7 @@ impl PublicGroups {
             latest_message_index: summary.latest_message_index,
             participant_count: summary.participant_count,
             is_public: true,
+            messages_visible_to_non_members: false,
             frozen: None,
             events_ttl: summary.events_ttl,
             events_ttl_last_updated: summary.events_ttl_last_updated,