+++ backend/canisters/group_index/impl/src/updates/c2c_create_group.rs
@@ -33,6 +33,7 @@ async fn c2c_create_group(args: Args) -> Response {
         subtype: args.subtype.clone(),
         avatar: args.avatar,
         history_visible_to_new_joiners: args.history_visible_to_new_joiners,
+        messages_visible_to_non_members: args.messages_visible_to_non_members,
         permissions_v2: args.permissions_v2,
         events_ttl: args.events_ttl,
         gate: args.gate.clone(),