+++ backend/integration_tests/src/update_group_tests.rs
@@ -35,6 +35,7 @@ fn update_group_name_succeeds() {
             public: None,
             correlation_id: 0,
             gate: NoChange,
+            messages_visible_to_non_members: None,
         },
     );
 
@@ -78,6 +79,7 @@ fn change_casing_of_group_name_succeeds() {
             public: None,
             correlation_id: 0,
             gate: NoChange,
+            messages_visible_to_non_members: None,
         },
     );
 
@@ -124,6 +126,7 @@ fn update_group_privacy_succeeds() {
             public: Some(false),
             correlation_id: 0,
             gate: NoChange,
+            messages_visible_to_non_members: None,
         },
     );
 
@@ -172,6 +175,7 @@ fn make_private_group_public_succeeds() {
             gate: NoChange,
             public: Some(true),
             correlation_id: 0,
+            messages_visible_to_non_members: None,
         },
     );