+++ backend/canisters/group/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Change max channel name length from 25 to 40 chars ([#6138](https://github.com/open-chat-labs/open-chat/pull/6138))
+- Configure message visibility to non-members of public groups ([#6152](https://github.com/open-chat-labs/open-chat/pull/6152))
 
 ## [[2.0.1262](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1262-group)] - 2024-07-26